refactor: move unit tests to own directory
