+++ api/src/__tests__/unit/shifts/GetShiftById.test.js
@@ -1,5 +1,5 @@
-const { Builder } = require('../builder')
-const { GetShiftById } = require('../../domains/shifts/use-cases/_index')
+const { Builder } = require('../../builder')
+const { GetShiftById } = require('../../../domains/shifts/use-cases/_index')
 
 const builder = new Builder()
