+++ api/src/__tests__/unit/shifts/Unsignup.test.js
@@ -1,5 +1,5 @@
-const { Builder } = require('../builder')
-const { Unsignup } = require('../../domains/shifts/use-cases/_index')
+const { Builder } = require('../../builder')
+const { Unsignup } = require('../../../domains/shifts/use-cases/_index')
 
 const builder = new Builder()
