+++ api/src/__tests__/unit/teams/DeleteLeadershipKey.test.js
@@ -1,5 +1,5 @@
-const { Builder } = require('../builder')
-const { DeleteLeadershipKey } = require('../../domains/teams/use-cases/_index')
+const { Builder } = require('../../builder')
+const { DeleteLeadershipKey } = require('../../../domains/teams/use-cases/_index')
 
 const builder = new Builder()
