+++ api/src/__tests__/unit/teams/GetLeads.test.js
@@ -1,5 +1,5 @@
-const { Builder } = require('../builder')
-const { GetLeads } = require('../../domains/teams/use-cases/_index')
+const { Builder } = require('../../builder')
+const { GetLeads } = require('../../../domains/teams/use-cases/_index')
 
 const builder = new Builder()
