+++ api/src/__tests__/unit/teams/GetShifts.test.js
@@ -1,5 +1,5 @@
-const { Builder } = require('../builder')
-const { GetShifts } = require('../../domains/teams/use-cases/_index')
+const { Builder } = require('../../builder')
+const { GetShifts } = require('../../../domains/teams/use-cases/_index')
 
 const builder = new Builder()
