+++ api/src/__tests__/unit/teams/RemoveLead.test.js
@@ -1,5 +1,5 @@
-const { Builder } = require('../builder')
-const { RemoveLead } = require('../../domains/teams/use-cases/_index')
+const { Builder } = require('../../builder')
+const { RemoveLead } = require('../../../domains/teams/use-cases/_index')
 
 const builder = new Builder()
