+++ api/src/__tests__/unit/users/AddUser.test.js
@@ -1,5 +1,5 @@
-const { Builder } = require('../builder')
-const { AddUser } = require('../../domains/users/use-cases/_index')
+const { Builder } = require('../../builder')
+const { AddUser } = require('../../../domains/users/use-cases/_index')
 
 const builder = new Builder()
